+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.7] - 2025-04-18
+
+### Added
+- Added lazyLoad prop to renderItem function in MasonryGallery component
+- Added lazyLoad prop to renderItem function in GridGallery component
+- Added lazyLoad prop to renderItem function in JustifiedGallery component
+- Added lazyLoad prop to renderItem function in StackGallery component
+
 ## [0.2.6] - 2025-04-18
 
 ### Added